[explorer/frontend] feat: add frontend variant selection
Introduce the frontend variant boundary so shared API imports resolve through the selected platform while preserving the existing NEM behavior as the default.

Move the NEM API implementations behind variants/nem and keep the public api modules as compatibility wrappers. Add config/env fallback coverage and the app config bootstrap path needed by the variant selector.
