+++ explorer/frontend/config/index.js
@@ -1,23 +1,43 @@
-const publicAppConfig = {
-	NATIVE_MOSAIC_ID: process.env.NATIVE_MOSAIC_ID,
-	NATIVE_MOSAIC_TICKER: process.env.NATIVE_MOSAIC_TICKER,
-	NATIVE_MOSAIC_DIVISIBILITY: +process.env.NATIVE_MOSAIC_DIVISIBILITY,
-	BLOCKCHAIN_UNWIND_LIMIT: +process.env.BLOCKCHAIN_UNWIND_LIMIT,
-	REQUEST_TIMEOUT: +process.env.REQUEST_TIMEOUT,
-	API_BASE_URL: process.env.API_BASE_URL,
-	SUPERNODE_API_URL: process.env.SUPERNODE_API_URL,
-	NODELIST_URL: process.env.NODELIST_URL,
-	MARKET_DATA_URL: process.env.MARKET_DATA_URL,
-	HISTORICAL_PRICE_URL: process.env.HISTORICAL_PRICE_URL,
-	SOCIAL_URL_TWITTER: process.env.SOCIAL_URL_TWITTER,
-	SOCIAL_URL_GITHUB: process.env.SOCIAL_URL_GITHUB,
-	SOCIAL_URL_DISCORD: process.env.SOCIAL_URL_DISCORD,
-	FOOTER_URL_DOCS: process.env.FOOTER_URL_DOCS,
-	FOOTER_URL_TECHNICAL_REFERENCE: process.env.FOOTER_URL_TECHNICAL_REFERENCE,
-	FOOTER_URL_FAUCET: process.env.FOOTER_URL_FAUCET,
-	FOOTER_URL_SUPERNODE_PROGRAM: process.env.FOOTER_URL_SUPERNODE_PROGRAM
+const env = name => process.env[`NEXT_PUBLIC_${name}`] || process.env[name];
+
+const platform = env('PLATFORM') || 'nem';
+
+const getVariantConfig = () => {
+	try {
+		// eslint-disable-next-line import/no-dynamic-require, global-require
+		return require(`../variants/${platform}/config`).config || {};
+	} catch {
+		return {};
+	}
 };
 
-const isClientSide = typeof window !== 'undefined';
+export const createPublicAppConfig = () => ({
+	PLATFORM: platform,
+	NATIVE_MOSAIC_ID: env('NATIVE_MOSAIC_ID'),
+	NATIVE_MOSAIC_TICKER: env('NATIVE_MOSAIC_TICKER'),
+	NATIVE_MOSAIC_DIVISIBILITY: +env('NATIVE_MOSAIC_DIVISIBILITY'),
+	BLOCKCHAIN_UNWIND_LIMIT: +env('BLOCKCHAIN_UNWIND_LIMIT'),
+	REQUEST_TIMEOUT: +env('REQUEST_TIMEOUT'),
+	API_BASE_URL: env('API_BASE_URL'),
+	SUPERNODE_API_URL: env('SUPERNODE_API_URL'),
+	NODELIST_URL: env('NODELIST_URL'),
+	MARKET_DATA_URL: env('MARKET_DATA_URL'),
+	HISTORICAL_PRICE_URL: env('HISTORICAL_PRICE_URL'),
+	SOCIAL_URL_TWITTER: env('SOCIAL_URL_TWITTER'),
+	SOCIAL_URL_GITHUB: env('SOCIAL_URL_GITHUB'),
+	SOCIAL_URL_DISCORD: env('SOCIAL_URL_DISCORD'),
+	FOOTER_URL_DOCS: env('FOOTER_URL_DOCS'),
+	FOOTER_URL_TECHNICAL_REFERENCE: env('FOOTER_URL_TECHNICAL_REFERENCE'),
+	FOOTER_URL_FAUCET: env('FOOTER_URL_FAUCET'),
+	FOOTER_URL_SUPERNODE_PROGRAM: env('FOOTER_URL_SUPERNODE_PROGRAM'),
+	...getVariantConfig()
+});
 
-export default isClientSide ? window.appConfig : publicAppConfig;
+export const serializeAppConfig = appConfig => JSON.stringify(appConfig).replace(/</g, '\\u003c');
+
+const publicAppConfig = createPublicAppConfig();
+
+const isClientSide = typeof window !== 'undefined';
+const getClientAppConfig = () => window.appConfig || window['__NEXT_DATA__']?.props?.appConfig || publicAppConfig;
+
+export default isClientSide ? getClientAppConfig() : publicAppConfig;
