+++ explorer/frontend/variants/nem/DocumentHead.jsx
@@ -1,0 +1,29 @@
+/* eslint-disable @next/next/no-page-custom-font */
+export const DocumentHead = () => (
+	<>
+		<meta charSet="utf-8" />
+		<link rel="icon" href="/favicon.ico" />
+		<meta name="format-detection" content="telephone=no" />
+		<meta name="theme-color" content="#F5F9FB" />
+		<meta name="description" content="NEM Block Explorer" />
+		<link rel="apple-touch-icon" href="/logo192.png" />
+		<link rel="manifest" href="/manifest.json" />
+		<link rel="preconnect" href="https://fonts.googleapis.com" />
+		<link
+			// eslint-disable-next-line max-len
+			href="https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,wght@0,200;0,300;0,400;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,600;1,700;1,800;1,900&display=swap"
+			rel="stylesheet"
+		/>
+		<link href="https://fonts.googleapis.com/css2?family=Abhaya+Libre:wght@400;500;600;700;800&display=swap" rel="stylesheet" />
+		<link
+			href="https://fonts.googleapis.com/css2?family=Courier+Prime:ital,wght@0,400;0,700;1,400;1,700&display=swap"
+			rel="stylesheet"
+		/>
+		<link
+			// eslint-disable-next-line max-len
+			href="https://fonts.googleapis.com/css2?family=Alegreya:ital,wght@0,400;0,500;0,600;0,700;0,800;0,900;1,400;1,500;1,600;1,700;1,800;1,900&display=swap"
+			rel="stylesheet"
+		/>
+		<link href="https://fonts.googleapis.com/css2?family=PT+Serif&display=swap" rel="stylesheet" />
+	</>
+);
