+++ explorer/frontend/variants/symbol/DocumentHead.jsx
@@ -1,0 +1,11 @@
+export const DocumentHead = () => (
+	<>
+		<meta charSet="utf-8" />
+		<link rel="icon" href="/symbol/favicon.svg" />
+		<meta name="format-detection" content="telephone=no" />
+		<meta name="theme-color" content="#1B0A29" />
+		<meta name="description" content="Symbol Block Explorer" />
+		<link rel="apple-touch-icon" href="/symbol/logo.png" />
+		<link rel="manifest" href="/symbol/manifest.json" />
+	</>
+);
